+++ doc/sphinx/src/news.rst
@@ -33,6 +33,10 @@ Deprecated
 
 Fixed
 ^^^^^
+* Data
+
+  * The ``particle`` layout works on the ``global`` topology.
+
 * Legion backend
 
   * Unit tests work without Kokkos but with OpenMP.